refactor(hycu): move type in types to be compliance with other uapp
ref: #NTNX-1
