+++ packages/manager/apps/hycu/src/mocks/licenseHycu/licenseHycu.data.ts
@@ -1,4 +1,4 @@
-import { IHycuDetails, LicenseStatus } from '@/type/hycu.details.interface';
+import { IHycuDetails, LicenseStatus } from '@/types/hycu.details.interface';
 
 export const licensesHycu: IHycuDetails[] = [
   {
