+++ packages/manager/apps/hycu/src/mocks/licenseHycu/licenseHycu.ts
@@ -1,4 +1,4 @@
-import { LicenseStatus } from '@/type/hycu.details.interface';
+import { LicenseStatus } from '@/types/hycu.details.interface';
 import { Handler } from '../../../../../../../playwright-helpers';
 import { licensesHycu } from './licenseHycu.data';
