+++ packages/manager/apps/hycu/src/pages/dashboard/general-information/ShortcutsTile/ShortcutsTile.spec.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 import { screen, waitFor } from '@testing-library/react';
 import { renderTestApp } from '@/utils/tests/renderTestApp';
 import { labels } from '@/utils/tests/init.i18n';
-import { LicenseStatus } from '@/type/hycu.details.interface';
+import { LicenseStatus } from '@/types/hycu.details.interface';
 
 describe('License Hycu shortcuts tile for dashboard test suite', () => {
   it('should show links of service to activate', async () => {
