+++ packages/manager/apps/hycu/src/pages/listing/Listing.page.tsx
@@ -28,7 +28,7 @@ import {
   ODS_BUTTON_SIZE,
 } from '@ovhcloud/ods-components';
 import HYCU_CONFIG from '@/hycu.config';
-import { IHycuDetails } from '@/type/hycu.details.interface';
+import { IHycuDetails } from '@/types/hycu.details.interface';
 
 import { urls, subRoutes } from '@/routes/routes.constant';
 import HycuActionMenu from './menu/HycuActionMenu.component';
