+++ packages/manager/apps/hycu/src/types/hycu.details.interface.ts
